Refactor to use theLoop
